feat: use `puppeteer-core` instead of `puppeteer`
`puppeteer-core` package is a version of `puppeteer` that doesn't download Chromium by default.
It looks like chromium is not directly used here as it was only a `peerDependencies` or an `optionalDependencies`.
So I moved to `puppeteer-core` to be able to compile TS sources of this addon without downloading Chromium at every `yarn install`